Initial commit: Start with GPLv3 license
As some pieces of code may by copied (either partially or completely)
directly from Moodle Jitsi and BigBlueButton plugins, and they both
are licensed under GPLv3, so the need to do so is obvious (respecting
the GPLv3 license itself).

By the way, this does not necessarily mean the source must be open-
source, although it is obvious for anyone knowing the GPL licenses
themselves.