Move InstanceDataForm back to mod_gharar_mod_form in mod_form.php
The parent class, i.e. moodleform_mod, extracts the module name
from the name of class using regex, and thus it sounds impossible
to have a class not in the form of mod_xx_mod_form. This is the
main reason of the change.

Also, the globals (including $CFG) required in the parent clas
makes the PSR4-based class definitions a bit weird.