Use a hacky workaround for making roles_can_view_recordings non-nullable
Because the column is a JSON string, and represents an array, it
should not be null, otherwise we might face future problems. Also,
text fields cannot have default values. So, instead of making it
nullable, first add it as nullable, but then put default values for
all records, and then change the field notnull property to false.
Well, I like the workaround pretty much. :)