Rename ToBeCreated* stuff to Possible*
A ToBeCreatedRoom, for example, may or may not exist. In the case
of being existed, it won't be created. So, the name PossibleRoom
better reflects this. Plus, somewhere we may need this more
general concept when dealing with stuff that its availability is
unknonw (e.g. whether someone is a member of a room or not).